refactor(plans): split oversized plan and module UI components
Break up timeline, pending-generation, create-plan, module header, and
module lessons surfaces into adjacent display and state modules so each
route keeps a thin orchestration shell. Preserve public imports and
behavior while applying small review follow-ups for copy, list keys,
attempt display consistency, and exhaustive module status styling.
