+++ src/app/(app)/plans/[id]/components/TimelineModuleCard.tsx
@@ -1,24 +1,22 @@
 'use client';
 
-import {
-  ArrowRight,
-  CheckCircle2,
-  ChevronRight,
-  ExternalLink,
-  Lock,
-} from 'lucide-react';
+import { ArrowRight, ChevronRight } from 'lucide-react';
 import Link from 'next/link';
 import type { JSX } from 'react';
-import { getResourceIcon } from '@/app/(app)/plans/resource-display';
-import { UpdateTaskStatusButton } from '@/app/(app)/plans/[id]/components/UpdateTaskStatusButton';
 import { AccordionContent, AccordionItem } from '@/components/ui/accordion';
 import { Button } from '@/components/ui/button';
-import { formatMinutes } from '@/features/plans/formatters';
 import { cn } from '@/lib/utils';
 import type { ClientTask } from '@/shared/types/client.types';
 import type { ProgressStatus } from '@/shared/types/db.types';
+import { TimelineModuleMarker } from './TimelineModuleMarker';
+import { TimelineTaskList } from './TimelineTaskList';
+import {
+  getCardClassName,
+  getTitleClassName,
+  getWeekBadgeClassName,
+} from './timeline-module-card-styles';
 
-type ModuleStatus = 'completed' | 'active' | 'locked';
+export type ModuleStatus = 'completed' | 'active' | 'locked';
 
 export interface TimelineModule {
   id: string;
@@ -38,187 +36,6 @@ interface TimelineModuleCardProps {
   statuses: Partial<Record<string, ProgressStatus>>;
   onModuleToggle: (moduleId: string) => void;
   onTaskStatusChange: (taskId: string, nextStatus: ProgressStatus) => void;
-}
-
-function getMarkerClassName(status: ModuleStatus): string {
-  if (status === 'completed') return 'border-success text-success';
-  if (status === 'active') {
-    return 'scale-110 border-primary text-primary shadow-[0_0_12px_hsl(var(--primary)/0.4)]';
-  }
-  return 'border-stone-300 text-stone-300 dark:border-stone-600 dark:text-stone-600';
-}
-
-function getCardClassName(status: ModuleStatus): string {
-  if (status === 'active') {
-    return 'border-primary/30 bg-white shadow-md dark:border-primary/50 dark:bg-stone-900';
-  }
-  if (status === 'locked') {
-    return 'border-stone-200 bg-stone-50 dark:border-stone-800 dark:bg-stone-900/70';
-  }
-  return 'border-stone-100 bg-white shadow-sm dark:border-stone-800 dark:bg-stone-900';
-}
-
-function getWeekBadgeClassName(status: ModuleStatus): string {
-  if (status === 'active') {
-    return 'bg-primary/10 text-primary dark:bg-primary/20 dark:text-primary';
-  }
-  if (status === 'completed') {
-    return 'bg-success/15 text-success dark:bg-success/25 dark:text-success-foreground';
-  }
-  return 'bg-stone-100 text-stone-500 dark:bg-stone-800 dark:text-stone-400';
-}
-
-function getTitleClassName(status: ModuleStatus): string {
-  if (status === 'active') return 'text-stone-900 dark:text-stone-100';
-  if (status === 'locked') return 'text-stone-600 dark:text-stone-400';
-  return 'text-stone-700 dark:text-stone-300';
-}
-
-function TimelineModuleMarker({ status }: { status: ModuleStatus }) {
-  return (
-    <div
-      className={cn(
-        'z-10 flex h-6 w-6 items-center justify-center rounded-full border-[3px] bg-white transition-all duration-500 ease-out dark:bg-stone-900',
-        getMarkerClassName(status),
-      )}
-    >
-      {status === 'completed' && (
-        <CheckCircle2 size={14} className="fill-green-100" />
-      )}
-      {status === 'active' && (
-        <div className="h-2 w-2 animate-pulse rounded-full bg-primary" />
-      )}
-      {status === 'locked' && <Lock size={10} />}
-    </div>
-  );
-}
-
-function TimelineResourceLink({
-  resource,
-}: {
-  resource: NonNullable<ClientTask['resources']>[number];
-}) {
-  const Icon = getResourceIcon(resource.type);
-
-  return (
-    <Button
-      key={resource.id}
-      variant="outline"
-      asChild
-      className="h-auto max-w-full justify-start rounded-lg px-2.5 py-1.5 text-left text-xs whitespace-normal"
-    >
-      <a href={resource.url} target="_blank" rel="noopener noreferrer">
-        <Icon size={14} className="shrink-0" />
-        <span className="wrap-break-word">{resource.title}</span>
-        <ExternalLink size={12} className="shrink-0 opacity-50" />
-      </a>
-    </Button>
-  );
-}
-
-function TimelineTaskCard({
-  task,
-  status,
-  onTaskStatusChange,
-}: {
-  task: ClientTask;
-  status: ProgressStatus;
-  onTaskStatusChange: (taskId: string, nextStatus: ProgressStatus) => void;
-}) {
-  const isCompleted = status === 'completed';
-  const resources = task.resources ?? [];
-
-  return (
-    <div
-      className={cn(
-        'rounded-2xl border p-4 transition-colors',
-        isCompleted
-          ? 'border-success/30 bg-success/5 dark:border-success/30 dark:bg-success/10'
-          : 'border-stone-100 bg-stone-50/50 hover:border-primary/30 dark:border-stone-800 dark:bg-stone-800/50 dark:hover:border-primary/50',
-      )}
-    >
-      <div className="flex h-full flex-col gap-3 sm:flex-row sm:items-center">
-        <div className="flex shrink-0 items-center">
-          <CheckCircle2
-            size={18}
-            className={cn(
-              isCompleted
-                ? 'fill-success/20 text-success dark:text-success'
-                : 'text-stone-300 dark:text-stone-600',
-            )}
-          />
-        </div>
-        <div className="flex min-w-0 flex-1 flex-col items-start justify-center">
-          <div className="flex flex-wrap items-center gap-2">
-            <p
-              className={cn(
-                'font-medium wrap-break-word',
-                isCompleted
-                  ? 'text-success dark:text-success'
-                  : 'text-stone-800 dark:text-stone-200',
-              )}
-            >
-              {task.title}
-            </p>
-            <span className="text-xs text-stone-400 dark:text-stone-500">
-              {formatMinutes(task.estimatedMinutes)}
-            </span>
-          </div>
-          {task.description && (
-            <p className="mt-1 text-sm wrap-break-word text-stone-500 dark:text-stone-400">
-              {task.description}
-            </p>
-          )}
-        </div>
-        <div className="flex shrink-0 items-center self-end sm:self-auto">
-          <UpdateTaskStatusButton
-            taskId={task.id}
-            status={status}
-            onStatusChange={onTaskStatusChange}
-          />
-        </div>
-      </div>
-
-      {resources.length > 0 && (
-        <div className="mt-3 ml-0 flex flex-wrap gap-2 sm:ml-6">
-          {resources.map((resource) => (
-            <TimelineResourceLink key={resource.id} resource={resource} />
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function TimelineTaskList({
-  module,
-  statuses,
-  onTaskStatusChange,
-}: {
-  module: TimelineModule;
-  statuses: Partial<Record<string, ProgressStatus>>;
-  onTaskStatusChange: (taskId: string, nextStatus: ProgressStatus) => void;
-}) {
-  if (module.tasks.length === 0) {
-    return (
-      <p className="text-sm text-stone-400 dark:text-stone-500">
-        No tasks in this module.
-      </p>
-    );
-  }
-
-  return (
-    <div className="space-y-3">
-      {module.tasks.map((task) => (
-        <TimelineTaskCard
-          key={task.id}
-          task={task}
-          status={statuses[task.id] ?? 'not_started'}
-          onTaskStatusChange={onTaskStatusChange}
-        />
-      ))}
-    </div>
-  );
 }
 
 export function TimelineModuleCard({
